+++ packages/cached_query/lib/cached_query.dart
@@ -5,5 +5,6 @@ export 'src/default_query_config.dart';
 export 'src/infinite_query_state.dart';
 export 'src/mutation.dart';
 export 'src/query_config.dart';
+export 'src/query_observer.dart';
 export 'src/query_state.dart';
 export 'src/storage_interface.dart';